function/requirements.txt
google-cloud-datastore==1.13.2